% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clusters_gpt.R
\name{annotate_clusters_gpt}
\alias{annotate_clusters_gpt}
\title{Annotate clusters using GPT and embed cell types into Seurat object}
\usage{
annotate_clusters_gpt(
  obj,
  markers_df,
  tissuename = "mouse liver",
  model = "gpt-4",
  output_csv = "Annotation.csv"
)
}
\arguments{
\item{obj}{Seurat object}

\item{markers_df}{Marker dataframe (e.g. output from FindAllMarkers)}

\item{tissuename}{Character. Tissue description (default = "mouse liver")}

\item{model}{Character. GPT model name (default = "gpt-4")}

\item{output_csv}{Character. Output CSV filename}
}
\value{
Seurat object with renamed identities and celltype column
}
\description{
Annotate clusters using GPT and embed cell types into Seurat object
}
