% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_spatial_regions.R
\name{annotate_spatial_regions}
\alias{annotate_spatial_regions}
\title{Interactively annotate spatial regions in a Seurat object}
\usage{
annotate_spatial_regions(
  seurat_obj,
  image,
  region_names = c("Region1"),
  assay = "SCT"
)
}
\arguments{
\item{seurat_obj}{A Seurat object with spatial data}

\item{image}{Name of spatial image (e.g. "condition2a")}

\item{region_names}{Character vector of region names to annotate}

\item{assay}{Assay to use for expression data (default = "SCT")}
}
\value{
Seurat object with new metadata column "region"
}
\description{
Interactively annotate spatial regions in a Seurat object
}
