% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_heatmap.R
\name{biomarker_heatmap}
\alias{biomarker_heatmap}
\title{Generate Heatmap from Marker Table}
\usage{
biomarker_heatmap(obj, marker_table)
}
\arguments{
\item{obj}{A Seurat object}

\item{marker_table}{A data frame containing marker genes}
}
\value{
A ggplot heatmap object
}
\description{
Generate Heatmap from Marker Table
}
