% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_celltype_composition.R
\name{calculate_celltype_composition}
\alias{calculate_celltype_composition}
\title{Prepare average percent cell type composition across groups}
\usage{
calculate_celltype_composition(
  seurat_obj,
  selected_celltypes = NULL,
  sample_col = "orig.ident",
  celltype_col = "celltype",
  group_col = "group"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{selected_celltypes}{Optional character vector of cell types to keep (default = NULL)}

\item{sample_col}{Metadata column indicating sample/replicate ID (default = "orig.ident")}

\item{celltype_col}{Metadata column indicating cell type (default = "celltype")}

\item{group_col}{Metadata column indicating experimental group (default = "group")}
}
\value{
A data frame containing group, cell type, average cell count,
and percent composition (ordered by variability)
}
\description{
Computes the average number of cells per sample within each group,
converts to percent within each cell type, and orders cell types
by variability across groups.
}
