% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cluster_biomarkers.R
\name{identify_cluster_biomarkers}
\alias{identify_cluster_biomarkers}
\title{Identify Cluster Biomarkers}
\usage{
identify_cluster_biomarkers(obj, only.pos = TRUE)
}
\arguments{
\item{obj}{A Seurat object}

\item{only.pos}{Logical; return only positive markers (default = TRUE)}

\item{assay}{Assay to use (default = DefaultAssay(obj))}
}
\value{
Data frame of all marker results
}
\description{
Identify Cluster Biomarkers
}
