% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pseudobulk.R
\name{perform_pseudobulk}
\alias{perform_pseudobulk}
\title{Perform pseudobulk aggregation from a Seurat object (raw counts)}
\usage{
perform_pseudobulk(
  seurat_obj,
  condition_col = "group",
  replicate_col = "orig.ident",
  celltype_col = "celltype",
  assay = "Spatial",
  return_seurat = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{condition_col}{Metadata column for condition (default = "group")}

\item{replicate_col}{Metadata column for replicate/sample (default = "orig.ident")}

\item{celltype_col}{Metadata column for cell type (default = "celltype")}

\item{assay}{Assay to use for pseudobulk (default = "RNA" raw counts)}

\item{return_seurat}{Logical; return Seurat object (default = TRUE)}
}
\value{
Pseudobulk Seurat object with \code{celltype_condition_replicate} as active identity
}
\description{
Perform pseudobulk aggregation from a Seurat object (raw counts)
}
