% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic_volcano.R
\name{plot_basic_volcano}
\alias{plot_basic_volcano}
\title{Generate Basic Volcano Plot with Gene Labels}
\usage{
plot_basic_volcano(
  deg_table,
  logfc_col = "avg_log2FC",
  padj_col = "p_val_adj",
  logfc_threshold = 0.5,
  padj_threshold = 0.01,
  label_logfc_threshold = 2,
  label_logq_threshold = 3,
  top_n = 20,
  save_path = NULL
)
}
\arguments{
\item{deg_table}{DEG results data frame}

\item{logfc_col}{Column name for log2 fold change}

\item{padj_col}{Column name for adjusted p-value}

\item{logfc_threshold}{Log2FC cutoff for coloring}

\item{padj_threshold}{Adjusted p-value cutoff for coloring}

\item{label_logfc_threshold}{Log2FC cutoff for labeling genes}

\item{label_logq_threshold}{-log10(FDR) cutoff for labeling genes}

\item{top_n}{Number of top genes (by |log2FC|) to label}

\item{save_path}{Optional file path to save SVG (without extension)}
}
\value{
ggplot object
}
\description{
Generate Basic Volcano Plot with Gene Labels
}
