% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_violin.R
\name{plot_cluster_violin}
\alias{plot_cluster_violin}
\title{Plot Violin Plot for Selected Genes}
\usage{
plot_cluster_violin(
  obj,
  features,
  ncol = 2,
  file = NULL,
  width = 6.5,
  height = 5
)
}
\arguments{
\item{obj}{A Seurat object}

\item{features}{Character vector of gene names}

\item{ncol}{Number of columns (default = 2)}

\item{file}{Optional file name for SVG export (without extension)}

\item{width}{SVG width (default = 6.5)}

\item{height}{SVG height (default = 5)}
}
\value{
A ggplot object
}
\description{
Plot Violin Plot for Selected Genes
}
