% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_cnet.R
\name{plot_go_cnet}
\alias{plot_go_cnet}
\title{Perform GO enrichment analysis and return grouped circular cnetplot}
\usage{
plot_go_cnet(
  deg_table,
  logfc_col = "avg_log2FC",
  padj_col = "p_val_adj",
  logfc_threshold = 0.5,
  padj_threshold = 0.05,
  orgdb = org.Mm.eg.db::org.Mm.eg.db,
  ont = "ALL",
  show_category = 5,
  circular = TRUE
)
}
\arguments{
\item{deg_table}{DEG results data frame}

\item{logfc_col}{Column name for log2 fold change}

\item{padj_col}{Column name for adjusted p-value}

\item{logfc_threshold}{Log2FC cutoff}

\item{padj_threshold}{Adjusted p-value cutoff}

\item{orgdb}{OrgDb annotation package}

\item{ont}{Ontology: "BP", "MF", or "CC"}

\item{show_category}{Number of GO categories to display (recommended 4–6)}

\item{circular}{Logical; circular layout}
}
\value{
ggplot object
}
\description{
Perform GO enrichment analysis and return grouped circular cnetplot
}
