% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_liver_qc_feature.R
\name{plot_liver_qc_feature}
\alias{plot_liver_qc_feature}
\title{Plot spatial feature-count QC for liver Seurat objects}
\usage{
plot_liver_qc_feature(seurat_obj, file = NULL, width = 6, height = 4.5)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{file}{Character or NULL. Output filename (without extension) for SVG export}

\item{width}{Numeric. Width of saved figure}

\item{height}{Numeric. Height of saved figure}
}
\value{
A patchwork plot object
}
\description{
Plot spatial feature-count QC for liver Seurat objects
}
