% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_regions.R
\name{plot_spatial_regions}
\alias{plot_spatial_regions}
\title{Plot spatial regions from a Seurat object with optional rotation}
\usage{
plot_spatial_regions(
  seurat_obj,
  group_by = "region",
  image,
  alpha = c(0.5, 0.5),
  pt_size_factor = 1.6,
  rotate_degree = 0
)
}
\arguments{
\item{seurat_obj}{A Seurat object with spatial data}

\item{group_by}{Metadata column to color by (default = "region")}

\item{image}{Name of spatial image}

\item{alpha}{Transparency vector (default = c(0.5, 0.5))}

\item{pt_size_factor}{Point size scaling factor (default = 1.6)}

\item{rotate_degree}{Degrees to rotate plot (0, 90, 180, 270)}
}
\value{
A ggplot object
}
\description{
Plot spatial regions from a Seurat object with optional rotation
}
