% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne_clustering.R
\name{plot_tsne_clustering}
\alias{plot_tsne_clustering}
\title{Plot t-SNE and Spatial Clusters}
\usage{
plot_tsne_clustering(object, label = TRUE, label.size = 3, combine = TRUE)
}
\arguments{
\item{object}{A processed Seurat object.}

\item{label}{Logical; whether to label clusters (default = TRUE).}

\item{label.size}{Size of spatial labels (default = 3).}

\item{combine}{Logical; whether to combine plots using patchwork (default = TRUE).}
}
\value{
A ggplot object (combined or individual plots)
}
\description{
Generates UMAP and spatial cluster plots and optionally combines them.
}
