% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_features.R
\name{plot_umap_features}
\alias{plot_umap_features}
\title{Plot Feature Expression on UMAP}
\usage{
plot_umap_features(obj, features, file = NULL, width = 5, height = 6)
}
\arguments{
\item{obj}{A Seurat object}

\item{features}{Character vector of gene names}

\item{file}{Optional file name for SVG export}

\item{width}{SVG width (default = 5)}

\item{height}{SVG height (default = 6)}
}
\value{
A ggplot object
}
\description{
Plot Feature Expression on UMAP
}
