% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_filter.R
\name{quantile_filter}
\alias{quantile_filter}
\title{Percentile-based feature filtering for Seurat objects}
\usage{
quantile_filter(
  obj,
  feature_col = "nFeature_Spatial",
  lower_pct = 0.01,
  upper_pct = 0.99
)
}
\arguments{
\item{obj}{A Seurat object}

\item{feature_col}{Metadata column to filter on}

\item{lower_pct}{Lower percentile (default = 0.01)}

\item{upper_pct}{Upper percentile (default = 0.99)}
}
\value{
A filtered Seurat object
}
\description{
Percentile-based feature filtering for Seurat objects
}
