% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dwls_annotation.R
\name{run_dwls_annotation}
\alias{run_dwls_annotation}
\title{Run DWLS deconvolution using a scRNA reference and return annotated Seurat object}
\usage{
run_dwls_annotation(
  spatial_obj,
  reference_obj,
  ref_celltype_col = "celltype",
  spatial_cluster_col = "seurat_clusters",
  spatial_assay = "Spatial",
  ref_assay = "SCT",
  downsample = 200,
  n_cell = 10
)
}
\arguments{
\item{spatial_obj}{Seurat spatial object (to annotate)}

\item{reference_obj}{Seurat scRNA-seq reference object}

\item{ref_celltype_col}{Column name in reference containing cell types}

\item{spatial_cluster_col}{Column in spatial object used for clusters}

\item{spatial_assay}{Assay name in spatial object (default = "Spatial")}

\item{ref_assay}{Assay name in reference object (default = "SCT")}

\item{downsample}{Number of cells per cell type for reference (default = 200)}

\item{n_cell}{Number of cells for DWLS (default = 10)}
}
\value{
Seurat object with DWLS deconvolution results
}
\description{
Run DWLS deconvolution using a scRNA reference and return annotated Seurat object
}
