% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rctd_annotation.R
\name{run_rctd_annotation}
\alias{run_rctd_annotation}
\title{Run RCTD deconvolution using a scRNA-seq reference}
\usage{
run_rctd_annotation(
  spatial_obj,
  reference_obj,
  ref_celltype_col = "celltype",
  ref_assay = "RNA",
  spatial_assay = "SCT",
  min_cells = 25,
  max_cores = 4,
  doublet_mode = "doublet"
)
}
\arguments{
\item{spatial_obj}{Seurat spatial object}

\item{reference_obj}{Seurat scRNA reference object}

\item{ref_celltype_col}{Column name in reference metadata containing cell types}

\item{ref_assay}{Assay name in reference (default = "RNA")}

\item{spatial_assay}{Assay name in spatial object (default = "SCT")}

\item{min_cells}{Minimum number of cells per reference cell type (default = 25)}

\item{max_cores}{Number of cores for RCTD (default = 4)}

\item{doublet_mode}{RCTD doublet mode ("doublet" or "full"; default = "doublet")}
}
\value{
Seurat object with RCTD results added to metadata
}
\description{
Run RCTD deconvolution using a scRNA-seq reference
}
