% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap_clustering.R
\name{run_umap_clustering}
\alias{run_umap_clustering}
\title{Run PCA, Clustering, and UMAP on a Seurat Object}
\usage{
run_umap_clustering(object, assay = "SCT", dims = 1:30, resolution = 1)
}
\arguments{
\item{object}{A Seurat object.}

\item{assay}{Assay to use for PCA (default = "SCT").}

\item{dims}{Dimensions to use for PCA and downstream steps (default = 1:30).}

\item{resolution}{Clustering resolution (default = 1).}
}
\value{
Processed Seurat object
}
\description{
Performs PCA, neighbor finding, clustering, and UMAP dimensional
reduction on a Seurat object.
}
