% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sample.R
\name{setup_sample}
\alias{setup_sample}
\title{Standardize Seurat metadata for downstream analyses}
\usage{
setup_sample(seurat_obj, condition)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{condition}{Character. Condition/sample name to assign}
}
\value{
A modified Seurat object
}
\description{
Standardize Seurat metadata for downstream analyses
}
